pyproject.toml
[project]
name = "video-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "ipykernel>=7.2.0",
    "jupyterlab>=4.5.5",
    "librosa>=0.11.0",
    "matplotlib>=3.10.8",
    "nltk>=3.9.3",
    "numpy>=2.4.2",
    "open-clip-torch>=3.3.0",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "transformers>=5.3.0",
    "wordcloud>=1.9.6",
    "yt-dlp>=2026.3.3",
]
